feat(groth16): point compression and reorganize module
