+++ Cargo.toml
@@ -26,7 +26,6 @@ ark-groth16 = { version = "0.3", optional=true }
 zeroize = "1.4"
 
 [features]
-default = ["r1cs"]
 r1cs = ["ark-r1cs-std", "ark-groth16"]
 
 [dev-dependencies]
